Make quad-bayer demosaic the default
